+++ backend/travelsite/src/main/java/com/teamapp/travelsite/Security/AuthController.java
@@ -1,10 +1,10 @@
 package com.teamapp.travelsite.Security;
 
 import com.teamapp.travelsite.Exception.BadRequestException;
-import com.teamapp.travelsite.Security.OAuthPayload.ApiResponse;
-import com.teamapp.travelsite.Security.OAuthPayload.AuthResponse;
-import com.teamapp.travelsite.Security.OAuthPayload.LoginRequest;
-import com.teamapp.travelsite.Security.OAuthPayload.SignUpRequest;
+import com.teamapp.travelsite.Security.OAuthBody.ApiResponse;
+import com.teamapp.travelsite.Security.OAuthBody.AuthResponse;
+import com.teamapp.travelsite.Security.OAuthBody.LoginRequest;
+import com.teamapp.travelsite.Security.OAuthBody.SignUpRequest;
 import com.teamapp.travelsite.User.User;
 import com.teamapp.travelsite.User.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
