+++ backend/travelsite/src/main/java/com/teamapp/travelsite/User/UserRepository.java
@@ -26,5 +26,4 @@ public interface UserRepository extends JpaRepository<User,Long> {
     Boolean existsByEmail(String email);
 
 
-
 }
