+++ frontend/react-social/src/user/login/Login.js
@@ -98,7 +98,7 @@ class LoginForm extends Component {
             Alert.success("You're successfully logged in!");
             this.props.history.push("/");
         }).catch(error => {
-            Alert.error((error && error.message) || 'Oops! Something went wrong. Please try again!');
+            Alert.error((error && error.message) || 'Oops! Something went wrong. Please try again');
         });
     }
